+++ docs/changelog.rst
@@ -10,11 +10,13 @@ CHANGELOG
 - Fix: do not prevent activity mappings overriding in subclasses of APIDAE Trek parser
 - Fix permissions bypass structure was always needed on accessibility attachments (#3396)
 - Fix default pictogram for mountainbike practice (it was blurry on mobile apps)
+- Fix: `delete=True` mode now works for APIDAE Trek parser
 
 
 **Improvements**
 
 - Add arguments loadsignage : sealing / manager (#3377)
+- Various minor improvements for APIDAE Trek parser
 
 
 2.97.4     (2023-03-09)